requirements.txt
vk_api
python-telegram-bot
pytz
requests
python-dotenv
retrying
prometheus-client
groq